feat: jwt token to the wallet daemon jrpc (#467)
Description
---
Add jwt token to wallet daemon.
When starting the webrtc, you need to get the token, the webrtc calls
then use this token for the jrpc server calls.
Currently the token is free to get (no real auth is going on). The
expiration and secret key are set in config.

How Has This Been Tested?
---
Not yet.

What process can a PR reviewer use to test or verify this change?
---
Can be tested by setting up the extension (including the signaling
server).

Breaking Changes
---
This will require dan wallet web ui to be updated.

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify