fix: don't send announce on address update (#464)
Temporarily disable announcing address updates to the network.

Currently we flood the network with announces if the peer's address is
updated, but this is not the correct solution. It should use the
complicated logic inside the PeerManager to decide when to forward
announces. The peer manager does a better job of checking identity
signatures.

Also, addresses don't change often

Created issue #463 to fix this issue in the long run