feat(wallet): use default account and fee for cli if not specified (#476)
Description
---
Change CLI commands to allow omitting the account or fee. If omitted,
the default account and fee is used.

Also changed all args for accounts to use either the name of the account
or component address.

Motivation and Context
---
The main idea here was to simplify the CLI args. Some CLI commands
expected a name of an account and others expected a component address.

How Has This Been Tested?
---
Tested all commands manually

What process can a PR reviewer use to test or verify this change?
---
Run a CLI command and omit the fee or account

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify