feat!: templates can specify the component_id (#499)
Description
---
* Updated the engine call for component creation so the `component_id`
can be specified
* Updated the template macro so the `component_id` is passed to the
engine
* Updated the `Account` template so the public key of the `owner_token`
is used as the `component_id`

Motivation and Context
---
On https://github.com/tari-project/tari-dan/pull/489 we refactored
component addresses so they are composed of `template_address` and a new
field `component_id` (which is a hash).

The purpose of this PR is to continue that line of work and:
* Allow templates to specify the `component_id` field
* Update the `Account` template so the corresponding public key of the
`owner_token` is used as the `component_id`

How Has This Been Tested?
---
Existing unit and integration tests pass

What process can a PR reviewer use to test or verify this change?
---
Create a new account component, deposit some funds and perform a
withdrawal

Breaking Changes
---
- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Both the `template_lib` crate and the `Account` builtin
template are updated