fix(vn): only return payload result from RPC/JRPC if finalized (#507)
Description
---
Only returns transaction result on transaction finality.
Increased the pacemaker timeout to 30s

Motivation and Context
---
The wallet awaits the transaction result after submission, which would
be returned even before consensus had committed it. This led the wallet
to accept the result (it shouldn't do that without validating the commit
phase QCs, but this is TODO) and update versions as if it succeeded but
if the transaction fails, the accounts/vaults/substates at `version + 1`
would not exist.

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---

Submit a transaction that fails. The walletd may see the result before
finality.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify