fix(walletd)!: free test commitment detectable by wallet, auto create account (#512)
Description
---
- make change outputs recoverable by always including encrypted value,
DH shared secret and public nonce in outputs
- allow confidential output to be specified directly by the wallet in
the CreateFreeTestCoins instruction
- `accounts create-free-test-coins MyAccount` command creates an account
if it does not exist (locally to the wallet)

Motivation and Context
---
Previously, the encrypted value and pub nonce (part of the derivation of
the shared DH mask) were optional and not used when producing change
outputs. This PR includes these in the change output so that the wallet
can recover them later.

CreateFreeTestCoins previously took a private mask. The wallet is not
able to open the resulting commitment. This PR changes that to take a
ConfidentialOutput and can be opened by the wallet.

In PR #510 we remove the ability to create accounts for free. This PR
enables accounts to be created with the free token deposit and fees paid
from that deposit in one transaction.

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
`cargo run --bin tari_dan_wallet_cli -- accounts create-free-test-coins
acc1` (`acc1` does not exist)

`cargo run --bin tari_dan_wallet_cli --  accounts get-balances acc1`

Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Change to free token instruction protobuf (old VNs wont
understand this instruction)