feat: adds event indexing over component addresses (#527)
Description
---

In previous work (#516, #500, #490, #478) we developed the tools for
both a tari indexer graphql service, templating events and adding
transaction receipts to the network substates (upon acceptance). The
goal of this PR is to bring forth the previous efforts together. Namely,
this PR aims at:

1. Add an event scanning interface, integrated with `SubstateManager` in
the tari indexer, that scans for events on the network and stores these
into the local indexer db for cache (whenever needed).
2. The above logic is integrated with the GraphQL interface, so that we
are to request events from a GraphQL service.
3. Refactored the `Event` interface (both on tari template lib and tari
indexer), so that component addresses are optional and we require both
`template_address` and `transaction_hash` fields.
4. We updated the tracker event formation, to automatically set the both
`transaction_hash` and `template_address` from its current state.
5. Refactored the enum `SubstateValue` so that deleted substates keep
track of which transaction hash they have been destroyed. In this way,
given a `ComponentAddress` and a `version` we can query directly the
network committee responsible for storing the substate containing the
`TransactionReceipt` associated with that same transaction. This greatly
improves the efficiency of the event scanning on the tari indexer, since
we don't have to query multiple committees to get the actual event
state.
6. Added `ConsensusConstants` to the `EpochManagerHandle`. This was done
in an initial version, to get the committee size at the level of
`SubstateManager` committee data to randomnize the process of querying
VNs within a committee. After a suggestion of @sdbondi, we can't do this
process in a more direct way. Therefore, we might as well remove this
step as well.

Motivation and Context
---

Updates the Tari Indexer service to be able to query the network to get
events associated with components (account creation, transfer, etc).
GraphQL integration is made.

How Has This Been Tested?
---
Added cucumber tests for querying events on the network, upon component
creation.

What process can a PR reviewer use to test or verify this change?
---
Run cucumber, create accounts and use the Tari Indexer to scan the
network for the corresponding events stored in the Indexer local db.
Manifests using the `emit_event` could also be used to further testing.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>