fix: remove unnecessary checked_sub and expect (#532)
Description
---
removes a checked_sub and expect which caused a panic

Motivation and Context
---
The Validator Node would panic when this checked_sub returns a negative.
I'm not sure exactly why it mismatched, but the base layer height was
39, so I assume some kind of edge case or perhaps a sync was happening.
Either way, the next line does a checked_sub and it doesn't look
necessary to prevent a negative number here.

How Has This Been Tested?
---
Validator Node would crash without this change when using dan_testing

What process can a PR reviewer use to test or verify this change?
---
Not really easy to reproduce

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify