feat: indexer auto-fills transaction inputs (#536)
Description
---
* Created a new `required_inputs` in the transaction meta, to hold the
inputs that the user/clients specifies.
* Clients and CLI populate the new `required_inputs` field.
* The indexer uses the autofill functionality in the
`submit_transaction` method (used by the wallet daemon) so all related
inputs are included.
* Moved substate decoding and transaction autofill functionality into
the `indexer_lib` to allow them to be reused in other applications in
the future.

Motivation and Context
---
Up until now, the client submitting a transaction into the network must
specify all inputs that are needed for consensus. This makes it
impractical to submit transactions as it can be quite difficult to know
all the inputs involved, specially on more complex templates.

So the goal of this PR is to **implement a input auto-filler in the
Indexer**, so clients only need to specify the higher level components
involved (i.e. accounts and the components they want to call) and the
indexer will automatically scan the network to include all related
substates (as well as the latests version of those).

For security reasons, we are creating a separated field in transactions
(`required_inputs`) which contains the inputs that the client requires
(typically the account address component with a specific version) and
it's read-only (i.e. included in the transaction hash). The indexer just
adds all related inputs into the existing `involved_objects` field that
is not included in the hash.

Note that there may be other auto-fill actions to be developed in the
future to have a seamlessly experience in transaction creation.

How Has This Been Tested?
---
- New cucumber scenario for transfers between accounts that exist
- Manually submitting a transaction

What process can a PR reviewer use to test or verify this change?
---
Submit a transaction thought the indexer, only specifying the component
addresses as inputs, and let the indexer figure out all other dependent
inputs

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted (new `required_inputs` field
in transactions)
- [ ] Other - Please specify