feat(walletui): transaction confirmation popup (#535)
Description
---
Adds a popup with a UI to confirm a transaction in the wallet.
Currently checks for the value: 'transaction://', but this can be
changed once we have the correct info.

Motivation and Context
---

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
The popup will trigger if you have 'transaction://' in your clipboard
(in Chrome / Safari).

Breaking Changes
---
x

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>