feat(consensus): add on vote receved and transaction executor (#568)
Description
---
feat(consensus): implements on vote received
feat(consensus): adds transaction executor
fix(consensus): correct safenode predicate in OnReceiveProposal

Motivation and Context
---
*OnReceiveVote*
Collects votes from replicas and generates a new high QC
On beat will trigger later and propose the next block containing the QC
for the previous block.

*TransactionExecutor*
Responsible for executing transactions in parallel when received from
the network and when missing transactions are requested in response to a
proposed block with unknown local transactions.

This component is not currently used in this PR as it will require
batched consensus integration with the validator node. This work is left
for a subsequent PR.

*safeNode predicate*
@stringhandler pointed out that I had the incorrect boolean logic in the
safeNode predicate. This has been corrected.

How Has This Been Tested?
---
Tests required

What process can a PR reviewer use to test or verify this change?
---
Tests to be created in subsequent PR

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify