feat: add wallet ui to the wallet daemon (#561)
Description
---
The server for the webui is added, but somehow it will not serve it. It
downloads the index.html also the .js but it will not run the js, so you
just see blank html.

Motivation and Context
---

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify