feat: dry run transactions in the indexer through the wallet (#576)
Description
---
* Created new dry run processor in the indexer
* Indexer now scans and stores templates (needed for executing the
transactions)
* Wallet daemon (and CLI) now properly handle the **dry run** option (up
until now it was ignored)
* The utility crate `tari_dan_app_utilities` now contains the
`TemplateProvider` as well as the `PayloadProcessor`. They are reused in
both `tari_validator_node` and `tari_indexer`
* The wallet daemon now does not store a transaction in the database if
it's a dry run
* Renamed in the wallet daemon `substate_provider.rs` to `network.rs`,
to better represent the logic of that module

Motivation and Context
---
**Dry run** transactions are transactions that are simulated, that means
the wasm is executed like a regular transaction but they are **not
committed into the network** (no consensus, no side effects, etc.). This
is useful for testing purposes and to call read-only methods without
paying fees.

Currently only the validator node supports dry run transactions. We need
a way to submit dry run transactions in the wallet daemon. A very easy
solution could be to simply make the indexer proxy the dry run request
to the validator node, but there is value on **executing the dry run
transaction in the indexer**:
* Read-only transactions will be very frequent. If they are executed in
the indexer we reduce the load on validator nodes (which are the ones
responsible for consensus).
* Simulating a transaction in the indexer will allow us in the future to
calculate a more precise input/output autofill

**TBD**: We may consider removing the dry run processor from the
validator node, as it's a user feature and VN's should only focus on
consensus. But this is debatable so I left it out of this PR.

How Has This Been Tested?
---
Manually submitted dry run transactions on the wallet CLI. For example:
```
cargo run --bin tari_dan_wallet_cli -- transactions submit --dry-run --num-outputs 2 call-function -a 100000 fe6b6a4ada71fd292d21235ee33289d3ed2a23f679334e65e1d7f9b314e790f2 mint
```
The wallet CLI only sends a JSON RPC request to the wallet daemon, so it
should be possible to do so from a web page as well.

What process can a PR reviewer use to test or verify this change?
---
Manually submit dry run transactions using the wallet CLI

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify