feat: wallet transactions page (#578)
Description
---
Adds a transactions page to the Dan Wallet
Currently still a work in progress to get feedback on what needs to be
displayed on the page

Motivation and Context
---
Adds the ability to view all transactions and inspect details

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
Run the app, the transactions will display at the bottom of the homepage
and also in the sidebar menu

Breaking Changes
---
x

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: stringhandler <mikethetike@tari.com>