fix(walletd): fix race condition causing transaction to be marked as invalid (#654)
Description
---
fix(walletd): fix race condition causing transaction to be marked as
invalid

Motivation and Context
---
Previously the mempool used to have no record of the transaction if it
failed validation. Now the mempool keeps track of this, which allows
wallets to query this (at least for a time). However, if the wallet
queries the transaction very soon after submitting it, the validator may
not have processed it yet and will return a not found result.

How Has This Been Tested?
---
Only tested manually

What process can a PR reviewer use to test or verify this change?
---
Submit transactions, test that they are never mistakenly marked as
invalid

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

Co-authored-by: stringhandler <mikethetike@tari.com>