fix(consensus)!: lock outputs objects (#659)
Description
---
fix(consensus)!: track and lock pending outputs to prevent potential
conflicts
refactor(transaction)!: remove filled_outputs from transaction type 
refactor(p2p messaging)!: add NewTransactionMessage type to DanMessage
with output_shards field
fix(mempool/consensus): do not propagate missing transactions that were
requested for the node
fix(mempool): take output_shards into account when deciding on involved
shards
fix(consensus): fix div by zero panic when calculating fees if involved
shards is (somehow) empty
fix(consensus): only unlock locked input/output substates when they were
originally locked
fix(mempool): create a list of output shards for the transaction after
execution instead of rebuilding that list multiple times
tests(consensus): fail quickly if a consensus error occurs by emitting a
Failure HotstuffEvent
tests(consensus): keep track of transactions in the test harness and
resend the original tx if requested

Motivation and Context
---
It is possible for multiple transactions to produce one or more
non-unique output shards. In the PR we lock outputs when preparing them
preventing any conflicting transactions from overwriting outputs.

Previously we "piggy-backed" the filled outputs on the Transaction type.
Filled output shards are the UP'd substates mapped into a ShardId and
are provided by an involved validator to a validator in one or more of
the output shards. Therefore these should not be settable by users. In
this PR I moved them to the output_shards field in the new
NewTransactionMessage struct.

In the test harness, we used default_fee and default_decision fields for
transactions. This is brittle and would cause unexpected behaviour if
testing fee disagreements or setting the fee to something else other
than the default.
In this PR I change the test harness to keep track of previous
transactions and provide them unchanged if requested.

How Has This Been Tested?
---
Consensus tests pass. It's a little difficult to test here because the
current behaviour is to not vote and from there the transaction is stuck
due to a bug. This would happen for all cases where there is local shard
disagreement so a fix is out of scope for this PR. I've written a test
case but it is currently not passing for this reason so it is commented
out.

What process can a PR reviewer use to test or verify this change?
---
Currently you can only test that consensus still works.

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify