docs: new web-wallet sequence diagram (#676)
Description
---
New `mermaid` sequence diagram with the full flow between a web user and
a wallet daemon

Motivation and Context
---
The flow of communications between a web and a Tari wallet have many
steps. We want a high level overview of it so we can make better
decisions regarding security and infrastructure

This PR adds a `mermaid` sequence diagram with the full flow between a
web user and a wallet daemon, from connection initialisation to messages
(e.g. transactions)

How Has This Been Tested?
---
This is only documentation, testing do not apply

What process can a PR reviewer use to test or verify this change?
---
Review that the diagram steps are accurate

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify